refactor code for func health events
