+++ Services/Implementations/HealthEventService.cs
@@ -3,6 +3,7 @@ using DTOs.HealthEventDTOs.Response;
 using Microsoft.Extensions.Logging;
 using Repositories.Interfaces;
 using Services.Commons;
+using Services.Mappers;
 using System.Data;
 
 namespace Services.Implementations
@@ -41,8 +42,7 @@ namespace Services.Implementations
                 var events = await _healthEventRepository.GetHealthEventsAsync(
                     pageNumber, pageSize, searchTerm, status, eventType, studentId, fromDate, toDate);
 
-                var eventDTOs = events.Select(MapToResponseDTO).ToList();
-                var result = CreatePagedResult(events, eventDTOs);
+                var result = events.ToPagedResponseDTO();
 
                 return ApiResult<PagedList<HealthEventResponseDTO>>.Success(
                     result, "Lấy danh sách sự kiện y tế thành công");
@@ -65,7 +65,7 @@ namespace Services.Implementations
                         new Exception("Không tìm thấy sự kiện y tế"));
                 }
 
-                var detailDTO = MapToDetailResponseDTO(healthEvent);
+                var detailDTO = HealthEventMapper.MapToDetailResponseDTO(healthEvent);
                 return ApiResult<HealthEventDetailResponseDTO>.Success(
                     detailDTO, "Lấy thông tin chi tiết sự kiện y tế thành công");
             }
@@ -89,7 +89,7 @@ namespace Services.Implementations
                             new Exception(validationResult.Message));
                     }
 
-                    var healthEvent = MapFromCreateRequest(request);
+                    var healthEvent = HealthEventMapper.MapFromCreateRequest(request);
 
                     // Tự động set trạng thái là Pending
                     healthEvent.EventStatus = EventStatus.Pending;
@@ -97,7 +97,7 @@ namespace Services.Implementations
 
                     var createdEvent = await CreateAsync(healthEvent);
 
-                    var eventDTO = MapToResponseDTO(createdEvent);
+                    var eventDTO = HealthEventMapper.MapToResponseDTO(createdEvent);
 
                     _logger.LogInformation("Health event created with status Pending: {EventId}", createdEvent.Id);
 
@@ -178,7 +178,7 @@ namespace Services.Implementations
 
                     // Lấy lại dữ liệu để trả về
                     var updatedEvent = await _healthEventRepository.GetByIdAsync(request.HealthEventId);
-                    var eventDTO = MapToResponseDTO(updatedEvent!);
+                    var eventDTO = HealthEventMapper.MapToResponseDTO(updatedEvent!);
 
                     var statusMessage = hasAnyTreatment && updatedEvent!.EventStatus == EventStatus.InProgress
                         ? "Thêm điều trị thành công và trạng thái được cập nhật thành InProgress"
@@ -232,7 +232,7 @@ namespace Services.Implementations
 
                     // Lấy lại dữ liệu để trả về
                     var updatedEvent = await _healthEventRepository.GetByIdAsync(request.HealthEventId);
-                    var eventDTO = MapToResponseDTO(updatedEvent!);
+                    var eventDTO = HealthEventMapper.MapToResponseDTO(updatedEvent!);
 
                     _logger.LogInformation("Health event {EventId} resolved successfully", request.HealthEventId);
 
@@ -402,8 +402,7 @@ namespace Services.Implementations
                 var events = await _healthEventRepository.GetSoftDeletedEventsAsync(
                     pageNumber, pageSize, searchTerm);
 
-                var eventDTOs = events.Select(MapToResponseDTO).ToList();
-                var result = CreatePagedResult(events, eventDTOs);
+                var result = events.ToPagedResponseDTO();
 
                 return ApiResult<PagedList<HealthEventResponseDTO>>.Success(
                     result, "Lấy danh sách sự kiện y tế đã xóa thành công");
@@ -490,6 +489,7 @@ namespace Services.Implementations
 
             return (true, "Valid");
         }
+
         private async Task<(bool IsSuccess, string Message)> ValidateEventMedicationsAsync(List<CreateEventMedicationRequest> medications)
         {
             foreach (var medication in medications)
@@ -517,7 +517,6 @@ namespace Services.Implementations
         private async Task<(bool IsSuccess, string Message)> ValidateSupplyUsagesAsync(List<CreateSupplyUsageRequest> supplies, Guid currentUserId)
         {
             // Kiểm tra current user có NurseProfile không
-            // You'll need a proper repository or method
             var nurseProfiles = await _unitOfWork.GetRepository<NurseProfile, Guid>()
                 .GetAllAsync(np => np.UserId == currentUserId);
             var nurseProfile = nurseProfiles.FirstOrDefault();
@@ -548,7 +547,6 @@ namespace Services.Implementations
 
             return (true, "Valid");
         }
-
 
         private async Task ProcessEventMedicationsAsync(Guid healthEventId, List<CreateEventMedicationRequest> medications,
             Guid currentUserId, DateTime vietnamTime)
@@ -622,37 +620,6 @@ namespace Services.Implementations
             }
         }
 
-        private async Task UpdateEventStatusBasedOnTreatmentAsync(Guid eventId, bool hasAnyTreatment, bool? isResolved, Guid currentUserId)
-        {
-            var healthEvent = await _healthEventRepository.GetByIdAsync(eventId);
-            if (healthEvent == null) return;
-
-            // Logic cập nhật trạng thái tự động
-            if (isResolved == true && healthEvent.EventStatus == EventStatus.InProgress)
-            {
-                // Chuyển sang Resolved
-                await _healthEventRepository.UpdateEventStatusAsync(eventId, EventStatus.Resolved, currentUserId);
-            }
-            else if (hasAnyTreatment && healthEvent.EventStatus == EventStatus.Pending)
-            {
-                // Có điều trị và đang Pending -> chuyển sang InProgress
-                await _healthEventRepository.UpdateEventStatusAsync(eventId, EventStatus.InProgress, currentUserId);
-            }
-        }
-
-        private static string GetStatusUpdateMessage(EventStatus currentStatus, bool hasAnyTreatment, bool? isResolved)
-        {
-            return currentStatus switch
-            {
-                EventStatus.Pending => "Cập nhật sự kiện y tế thành công (trạng thái: Pending)",
-                EventStatus.InProgress when hasAnyTreatment => "Thêm điều trị thành công và trạng thái được cập nhật thành InProgress",
-                EventStatus.InProgress => "Cập nhật sự kiện y tế thành công (trạng thái: InProgress)",
-                EventStatus.Resolved when isResolved == true => "Hoàn thành xử lý sự kiện y tế thành công (trạng thái: Resolved)",
-                EventStatus.Resolved => "Cập nhật sự kiện y tế thành công (trạng thái: Resolved)",
-                _ => "Cập nhật sự kiện y tế thành công"
-            };
-        }
-
         private static (bool isValid, string message) ValidateBatchInput(List<Guid> ids, string operation)
         {
             if (ids == null || !ids.Any())
@@ -699,101 +666,6 @@ namespace Services.Implementations
             {
                 return $"Không có sự kiện y tế nào được {operation}";
             }
-        }
-
-        private static HealthEvent MapFromCreateRequest(CreateHealthEventRequestDTO request)
-        {
-            return new HealthEvent
-            {
-                StudentId = request.StudentId,
-                EventCategory = request.EventCategory,
-                VaccinationRecordId = request.VaccinationRecordId,
-                EventType = request.EventType,
-                Description = request.Description,
-                OccurredAt = request.OccurredAt
-            };
-        }
-        private static HealthEventResponseDTO MapToResponseDTO(HealthEvent healthEvent)
-        {
-            return new HealthEventResponseDTO
-            {
-                Id = healthEvent.Id,
-                StudentId = healthEvent.StudentId,
-                StudentName = healthEvent.Student?.FullName ?? string.Empty,
-                EventCategory = healthEvent.EventCategory.ToString(),
-                VaccinationRecordId = healthEvent.VaccinationRecordId,
-                EventType = healthEvent.EventType.ToString(),
-                Description = healthEvent.Description,
-                OccurredAt = healthEvent.OccurredAt,
-                EventStatus = healthEvent.EventStatus.ToString(),
-                ReportedBy = healthEvent.ReportedUserId,
-                ReportedByName = healthEvent.ReportedUser?.FullName ?? string.Empty,
-                CreatedAt = healthEvent.CreatedAt,
-                UpdatedAt = healthEvent.UpdatedAt,
-                IsDeleted = healthEvent.IsDeleted,
-                TotalMedications = healthEvent.EventMedications?.Count ?? 0,
-                TotalSupplies = healthEvent.SupplyUsages?.Count ?? 0
-            };
-        }
-
-        private static HealthEventDetailResponseDTO MapToDetailResponseDTO(HealthEvent healthEvent)
-        {
-            var baseDto = MapToResponseDTO(healthEvent);
-
-            return new HealthEventDetailResponseDTO
-            {
-                Id = baseDto.Id,
-                StudentId = baseDto.StudentId,
-                StudentName = baseDto.StudentName,
-                EventCategory = baseDto.EventCategory,
-                VaccinationRecordId = baseDto.VaccinationRecordId,
-                EventType = baseDto.EventType,
-                Description = baseDto.Description,
-                OccurredAt = baseDto.OccurredAt,
-                EventStatus = baseDto.EventStatus,
-                ReportedBy = baseDto.ReportedBy,
-                ReportedByName = baseDto.ReportedByName,
-                CreatedAt = baseDto.CreatedAt,
-                UpdatedAt = baseDto.UpdatedAt,
-                IsDeleted = baseDto.IsDeleted,
-                TotalMedications = baseDto.TotalMedications,
-                TotalSupplies = baseDto.TotalSupplies,
-
-                Medications = healthEvent.EventMedications?.Select(em => new EventMedicationResponseDTO
-                {
-                    Id = em.Id,
-                    MedicationLotId = em.MedicationLotId,
-                    MedicationName = em.MedicationLot?.Medication?.Name ?? string.Empty,
-                    LotNumber = em.MedicationLot?.LotNumber ?? string.Empty,
-                    Quantity = em.Quantity,
-                    UsedAt = em.CreatedAt
-                }).ToList() ?? new List<EventMedicationResponseDTO>(),
-
-                Supplies = healthEvent.SupplyUsages?.Select(su => new SupplyUsageResponseDTO
-                {
-                    Id = su.Id,
-                    HealthEventId = su.HealthEventId,
-                    MedicalSupplyLotId = su.MedicalSupplyLotId,
-                    MedicalSupplyName = su.MedicalSupplyLot?.MedicalSupply?.Name ?? string.Empty,
-                    LotNumber = su.MedicalSupplyLot?.LotNumber ?? string.Empty,
-                    QuantityUsed = su.QuantityUsed,
-                    NurseProfileId = su.NurseProfileId,
-                    NurseName = su.UsedByNurse?.User?.FullName ?? string.Empty,
-                    Notes = su.Notes,
-                    CreatedAt = su.CreatedAt
-                }).ToList() ?? new List<SupplyUsageResponseDTO>()
-            };
-        }
-
-        private static PagedList<HealthEventResponseDTO> CreatePagedResult(
-            PagedList<HealthEvent> sourcePaged,
-            List<HealthEventResponseDTO> mappedItems)
-        {
-            return new PagedList<HealthEventResponseDTO>(
-                mappedItems,
-                sourcePaged.MetaData.TotalCount,
-                sourcePaged.MetaData.CurrentPage,
-                sourcePaged.MetaData.PageSize);
         }
 
         #endregion
